+++ src/plone/folder/ordered.py
@@ -10,13 +10,18 @@ from Products.BTreeFolder2.BTreeFolder2 import _marker
 from Products.BTreeFolder2.BTreeFolder2 import BTreeFolder2Base
 from Products.CMFCore.permissions import ModifyPortalContent
 from Products.CMFCore.PortalFolder import PortalFolderBase
-from Products.ZCatalog.Lazy import LazyMap
 from zope.annotation.interfaces import IAttributeAnnotatable
 from zope.component import getAdapter
 from zope.component import queryAdapter
 from zope.interface import implementer
+try:
+    from ZTUtils.Lazy import LazyMap
+except ImportError:
+    # bbb import for Zope2
+    from Products.ZCatalog.Lazy import LazyMap
 
 import pkg_resources
+
 
 HAS_ZSERVER = True
 try:
@@ -38,7 +43,7 @@ class OrderedBTreeFolderBase(BTreeFolder2Base):
 
     security = ClassSecurityInfo()
 
-    def __nonzero__(self):
+    def __bool__(self):
         """ a folder is something, even if it's empty """
         return True